refactor: drop dead vterm search/range API
Search, ScrollToLine, GetAllLines, VisibleLineRange, TotalLines and the rowToString helper in internal/vterm/search.go were exported but never called by any UI or app code (only tests exercised them), so Go's unused-symbol lint could never flag them. Delete the whole speculative search/visible-range cluster and its dedicated tests. The one live symbol in that file, MaxViewOffset (used by the sidebar/center wheel handlers), moves to vterm_scroll.go next to currentMaxViewOffset, with its tests relocated to vterm_scroll_test.go. Part of the quality stacked diff (stacked on main).
