+++ internal/vterm/vterm_widechar_test.go
@@ -48,28 +48,6 @@ func TestWideCharacterAtEndOfLine(t *testing.T) {
 	}
 }
 
-func TestRowToStringSkipsContinuationCells(t *testing.T) {
-	t.Parallel()
-	vt := New(10, 1)
-	vt.Write([]byte("你A"))
-
-	got := rowToString(vt.Screen[0])
-	if got != "你A" {
-		t.Fatalf("rowToString() = %q, want %q", got, "你A")
-	}
-}
-
-func TestSearchSkipsContinuationCells(t *testing.T) {
-	t.Parallel()
-	vt := New(10, 1)
-	vt.Write([]byte("你A"))
-
-	matches := vt.Search("你A")
-	if len(matches) != 1 || matches[0] != 0 {
-		t.Fatalf("Search() = %v, want [0]", matches)
-	}
-}
-
 func TestGetSelectedTextSkipsContinuationCells(t *testing.T) {
 	t.Parallel()
 	vt := New(10, 1)
